# Experimenting with  barrier which proved to be killing our performance